+++ PecuniaFinanceSystem1/src/main/java/com/capgemini/pecunia/bean/LoanDisbursal.java
@@ -7,7 +7,7 @@ import javax.persistence.Table;
 @Entity
 @Table(name="loandisbursal")
 public class LoanDisbursal {
-	private long accountId;
+	private int accountId;
 	private double loanAmount;
 	private int loanTenure;
 	private int creditScore;
@@ -17,10 +17,10 @@ public class LoanDisbursal {
 	private double emi;
 	@Id
 	private int loanId;
-	public long getAccountId() {
+	public int getAccountId() {
 		return accountId;
 	}
-	public void setAccountId(long accountId) {
+	public void setAccountId(int accountId) {
 		this.accountId = accountId;
 	}
 	public double getLoanAmount() {
